+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## v8.7.3
+* Added a new "Any Totem" condition that will check all totem slots.
+
 ### Bug Fixes
 * Fix #1742 - Errors related to improper escaping of user input for the suggestion list.